fix(admin/ui): Move MainSidebar Components to @northware/ui
Die Komponenten,die später für die Haupt Sidebar genutzt werden sollen wurden aus der Admin App in @northware/ui verschoben.
