+++ packages/ui/components/main-sidebar.tsx
@@ -1,0 +1,99 @@
+"use client";
+import { NavMain } from "@northware/ui/components/nav-main";
+import { NavUser } from "@northware/ui/components/nav-user";
+import {
+  Sidebar,
+  SidebarContent,
+  SidebarFooter,
+  SidebarHeader,
+  SidebarRail,
+} from "@northware/ui/components/sidebar";
+import { TeamSwitcher } from "@northware/ui/components/team-switcher";
+
+// This is sample data.
+const data = {
+  user: {
+    name: "shadcn",
+    email: "m@example.com",
+    avatar: "/avatars/shadcn.jpg",
+  },
+  navMain: [
+    {
+      title: "Home",
+      url: "/",
+    },
+    {
+      title: "Benutzerverwaltung",
+      url: "/users",
+      items: [
+        {
+          title: "Rollen",
+          url: "/users/roles",
+        },
+        {
+          title: "Berechtigungsschlüssel",
+          url: "/users/permission-keys",
+        },
+      ],
+    },
+    {
+      title: "HR Managment",
+      url: "#",
+      items: [
+        {
+          title: "Genesis",
+          url: "#",
+        },
+        {
+          title: "Explorer",
+          url: "#",
+        },
+        {
+          title: "Quantum",
+          url: "#",
+        },
+      ],
+    },
+    {
+      title: "Settings",
+      url: "#",
+      items: [
+        {
+          title: "General",
+          url: "#",
+        },
+        {
+          title: "Team",
+          url: "#",
+        },
+        {
+          title: "Billing",
+          url: "#",
+        },
+        {
+          title: "Limits",
+          url: "#",
+        },
+      ],
+    },
+  ],
+};
+
+export function MainSidebar({
+  ...props
+}: React.ComponentProps<typeof Sidebar>) {
+  return (
+    <Sidebar {...props} variant="inset" collapsible="offcanvas">
+      <SidebarHeader>
+        <TeamSwitcher />
+      </SidebarHeader>
+      <SidebarContent>
+        <NavMain items={data.navMain} />
+      </SidebarContent>
+      <SidebarFooter>
+        <NavUser user={data.user} />
+      </SidebarFooter>
+      <SidebarRail />
+    </Sidebar>
+  );
+}
