+++ packages/ui/components/team-switcher.tsx
@@ -1,0 +1,95 @@
+"use client";
+
+import { ChevronsUpDown } from "@northware/ui/icons/lucide";
+import * as React from "react";
+
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@northware/ui/components/dropdown-menu";
+import {
+  SidebarMenu,
+  SidebarMenuButton,
+  SidebarMenuItem,
+  useSidebar,
+} from "@northware/ui/components/sidebar";
+import {
+  AdminIcon,
+  CockpitIcon,
+  FinanceIcon,
+  TraderIcon,
+} from "@northware/ui/icons/output";
+
+const teams = [
+  {
+    name: "Northware Cockpit",
+    logo: CockpitIcon,
+  },
+  {
+    name: "Northware Finance",
+    logo: FinanceIcon,
+  },
+  {
+    name: "Northware Trader",
+    logo: TraderIcon,
+  },
+  {
+    name: "Admin Panel",
+    logo: AdminIcon,
+  },
+];
+
+export function TeamSwitcher() {
+  const { isMobile } = useSidebar();
+  const [activeTeam, setActiveTeam] = React.useState(teams[0]);
+
+  if (!activeTeam) {
+    return null;
+  }
+
+  return (
+    <SidebarMenu>
+      <SidebarMenuItem>
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <SidebarMenuButton
+              size="lg"
+              className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
+            >
+              <div className="flex aspect-square size-8 items-center justify-center rounded-lg text-sidebar-primary-foreground">
+                <activeTeam.logo className="" />
+              </div>
+              <div className="grid flex-1 text-left text-sm leading-tight">
+                <span className="truncate font-semibold">
+                  {activeTeam.name}
+                </span>
+              </div>
+              <ChevronsUpDown className="ml-auto" />
+            </SidebarMenuButton>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent
+            className="w-[--radix-dropdown-menu-trigger-width] min-w-56 rounded-lg"
+            align="start"
+            side={isMobile ? "bottom" : "right"}
+            sideOffset={4}
+          >
+            {teams.map((team) => (
+              <DropdownMenuItem
+                key={team.name}
+                onClick={() => setActiveTeam(team)}
+                className="gap-2 p-2"
+              >
+                <div className="flex aspect-square size-8 items-center justify-center rounded-sm border text-sidebar-primary-foreground">
+                  <team.logo className="" />
+                </div>
+                {team.name}
+              </DropdownMenuItem>
+            ))}
+          </DropdownMenuContent>
+        </DropdownMenu>
+      </SidebarMenuItem>
+    </SidebarMenu>
+  );
+}
